[#100] 체크포인트 구현
