+++ Assets/_MyAssets/CheckPoint.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 59253adb604613647943edf0ea52543a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
