+++ Assets/_MyAssets/Scenes/Workspace/Respawn.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 65fd0039f3e67f24ebe2b1ce31997f10
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
