+++ Assets/_MyAssets/Scenes/Workspace/Respawn/Respawn.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 38d42cc503c345f4d80cb54635cb951d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
