+++ Assets/_MyAssets/Scripts/Misc/CheckPointHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 87632d8f18d5dae4f98f80db4968d8c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
