+++ Assets/_MyAssets/Scripts/SceneManagers/WorkSpace.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 29448e0bd2b48464eab3e6ff8540abbd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
