+++ Assets/_MyAssets/Scripts/SceneManagers/WorkSpace/CheckPointSceneManager.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CheckPointSceneManager : SceneManagerBase
+{
+    protected override void Start()
+    {
+        base.Start();
+        FadeIn(0.5f);
+    }
+
+    protected override void Update()
+    {
+        base.Update();
+    }
+}
