+++ Assets/_MyAssets/Scripts/SceneManagers/WorkSpace/CheckPointSceneManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ecc8bdab3fe05746b63f700133c323a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
